fix(interpeter): Shrink public API
